+++ src/ephemdir/__init__.py
@@ -45,7 +45,7 @@ from .core import (
     tempdir,
 )
 
-__version__ = "0.5.0"
+__version__ = "0.6.0"
 __author__ = "vindfjur"
 
 __all__ = [